split test files
